Support array as value for parsing paths.
Hiera don't support array as value for parse_string.
Check https://docs.puppetlabs.com/hiera/1/variables.html\#interpolation-tokens
This patch add an option to make the difference between string and array paths.
If it's an array simply create a path with each value in the array.
We can't get the entire scope but we can check the key.
For avoiding rewritring regex we add a key to the path.